major OSPF overhaul
- fix inter-area routing
- update Ospf*Route classes to support correct semantics
- introduce max-metric for:
  - router-lsa
  - stub networks
  - external networks
  - summary networks
- introduce point-to-point ospf interace option
  - uses stub network cost in certain cases
- add tests
- refactor data plane tests
- add ospf tests
