+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/OspfProcess.java
@@ -23,6 +23,14 @@ public class OspfProcess implements Serializable {
 
     private String _exportPolicy;
 
+    private Long _maxMetricExternalNetworks;
+
+    private Long _maxMetricStubNetworks;
+
+    private Long _maxMetricSummaryNetworks;
+
+    private Long _maxMetricTransitLinks;
+
     private Vrf _vrf;
 
     Builder(NetworkFactory networkFactory) {
@@ -36,11 +44,35 @@ public class OspfProcess implements Serializable {
         _vrf.setOspfProcess(ospfProcess);
       }
       ospfProcess.setExportPolicy(_exportPolicy);
+      ospfProcess.setMaxMetricExternalNetworks(_maxMetricExternalNetworks);
+      ospfProcess.setMaxMetricStubNetworks(_maxMetricStubNetworks);
+      ospfProcess.setMaxMetricSummaryNetworks(_maxMetricSummaryNetworks);
+      ospfProcess.setMaxMetricTransitLinks(_maxMetricTransitLinks);
       return ospfProcess;
     }
 
     public Builder setExportPolicy(RoutingPolicy exportPolicy) {
       _exportPolicy = exportPolicy != null ? exportPolicy.getName() : null;
+      return this;
+    }
+
+    public Builder setMaxMetricExternalNetworks(Long maxMetricExternalNetworks) {
+      _maxMetricExternalNetworks = maxMetricExternalNetworks;
+      return this;
+    }
+
+    public Builder setMaxMetricStubNetworks(Long maxMetricStubNetworks) {
+      _maxMetricStubNetworks = maxMetricStubNetworks;
+      return this;
+    }
+
+    public Builder setMaxMetricSummaryNetworks(Long maxMetricSummaryNetworks) {
+      _maxMetricSummaryNetworks = maxMetricSummaryNetworks;
+      return this;
+    }
+
+    public Builder setMaxMetricTransitLinks(Long maxMetricTransitLinks) {
+      _maxMetricTransitLinks = maxMetricTransitLinks;
       return this;
     }
 
@@ -52,6 +84,14 @@ public class OspfProcess implements Serializable {
 
   private static final int DEFAULT_CISCO_VLAN_OSPF_COST = 1;
 
+  private static final String PROP_MAX_METRIC_EXTERNAL_NETWORKS = "maxMetricExternalNetworks";
+
+  private static final String PROP_MAX_METRIC_STUB_NETWORKS = "maxMetricStubNetworks";
+
+  private static final String PROP_MAX_METRIC_SUMMARY_NETWORKS = "maxMetricSummaryNetworks";
+
+  private static final String PROP_MAX_METRIC_TRANSIT_LINKS = "maxMetricTransitLinks";
+
   private static final long serialVersionUID = 1L;
 
   private SortedMap<Long, OspfArea> _areas;
@@ -59,6 +99,14 @@ public class OspfProcess implements Serializable {
   private String _exportPolicy;
 
   private SortedSet<GeneratedRoute> _generatedRoutes;
+
+  private Long _maxMetricExternalNetworks;
+
+  private Long _maxMetricStubNetworks;
+
+  private Long _maxMetricSummaryNetworks;
+
+  private Long _maxMetricTransitLinks;
 
   private transient Map<Pair<Ip, Ip>, OspfNeighbor> _ospfNeighbors;
 
@@ -116,6 +164,26 @@ public class OspfProcess implements Serializable {
     return _generatedRoutes;
   }
 
+  @JsonProperty(PROP_MAX_METRIC_EXTERNAL_NETWORKS)
+  public Long getMaxMetricExternalNetworks() {
+    return _maxMetricExternalNetworks;
+  }
+
+  @JsonProperty(PROP_MAX_METRIC_STUB_NETWORKS)
+  public Long getMaxMetricStubNetworks() {
+    return _maxMetricStubNetworks;
+  }
+
+  @JsonProperty(PROP_MAX_METRIC_SUMMARY_NETWORKS)
+  public Long getMaxMetricSummaryNetworks() {
+    return _maxMetricSummaryNetworks;
+  }
+
+  @JsonProperty(PROP_MAX_METRIC_TRANSIT_LINKS)
+  public Long getMaxMetricTransitLinks() {
+    return _maxMetricTransitLinks;
+  }
+
   @JsonIgnore
   public Map<Pair<Ip, Ip>, OspfNeighbor> getOspfNeighbors() {
     return _ospfNeighbors;
@@ -157,6 +225,26 @@ public class OspfProcess implements Serializable {
 
   public void setGeneratedRoutes(SortedSet<GeneratedRoute> generatedRoutes) {
     _generatedRoutes = generatedRoutes;
+  }
+
+  @JsonProperty(PROP_MAX_METRIC_EXTERNAL_NETWORKS)
+  public void setMaxMetricExternalNetworks(Long maxMetricExternalNetworks) {
+    _maxMetricExternalNetworks = maxMetricExternalNetworks;
+  }
+
+  @JsonProperty(PROP_MAX_METRIC_STUB_NETWORKS)
+  public void setMaxMetricStubNetworks(Long maxMetricStubNetworks) {
+    _maxMetricStubNetworks = maxMetricStubNetworks;
+  }
+
+  @JsonProperty(PROP_MAX_METRIC_SUMMARY_NETWORKS)
+  public void setMaxMetricSummaryNetworks(Long maxMetricSummaryNetworks) {
+    _maxMetricSummaryNetworks = maxMetricSummaryNetworks;
+  }
+
+  @JsonProperty(PROP_MAX_METRIC_TRANSIT_LINKS)
+  public void setMaxMetricTransitLinks(Long maxMetricTransitLinks) {
+    _maxMetricTransitLinks = maxMetricTransitLinks;
   }
 
   @JsonIgnore
